realtychance_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
django-cors-headers==4.3.0
django-filter==23.3
Pillow==10.1.0